+++ ui/packages/app/grafana-datasource-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.33](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.32...@parca/grafana-datasource@0.0.33) (2022-12-12)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
 ## 0.0.32 (2022-12-10)
 
 **Note:** Version bump only for package @parca/grafana-datasource