+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.82](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.81...@parca/profile@0.16.82) (2022-12-12)
+
+**Note:** Version bump only for package @parca/profile
+
 ## 0.16.81 (2022-12-10)
 
 **Note:** Version bump only for package @parca/profile